Rename and simplify SystemTimingSupport
